Fix warnings.
In #785 the type of YR_MEMORY_BLOCK's "size" and "base" fields was changed from size_t to uint64_t. This changed introduced a number of warnings due to loss of precision when forcing the uint64_t into a size_t.

+++ libyara/include/yara/scan.h
@@ -43,7 +43,7 @@ int yr_scan_verify_match(
     YR_AC_MATCH* ac_match,
     const uint8_t* data,
     size_t data_size,
-    size_t data_base,
+    uint64_t data_base,
     size_t offset);
 
 #endif
